refactor(platforms): remove github provider
