+++ platforms/src/Github/index.ts
@@ -1,3 +1,2 @@
 export { GithubPlatform } from "./App-Bindings";
 export { ProviderConfig, PlatformDetails, providers } from "./Providers-config";
-export { GithubProvider } from "./Providers";
